backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
markdown==3.4.1
Pillow==9.3.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Markdown==3.3.4
